# docs/apidog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FundLedger API",
    "version": "0.1.0",
    "description": "Apidog-importable OpenAPI collection for the FundLedger multi-tenant SaaS API."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.fundledger.app",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/uploads/{id}/view": {
      "get": {
        "tags": ["Uploads"],
        "summary": "View uploaded attachment",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "responses": { "200": { "description": "Attachment binary content" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ProvisioningApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Billing/provisioning secret configured via PROVISIONING_API_KEYS."
      },
      "ProvisioningAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative form: Authorization: ApiKey <key>."
      }
    },
    "parameters": {
      "ProjectIdHeader": {
        "name": "X-Project-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Active project scope. If omitted, the session active project is used."
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    },
    "responses": {
...
      "Error": {
        "description": "Error response envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    },
...
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION" },
              "message": { "type": "string", "example": "Invalid request" },
              "fields": { "type": "object", "additionalProperties": true }
            }
          }
        }
      },
...
    }
  }
}